requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
pgvector
psycopg2-binary
pydantic
python-dotenv
streamlit
wikipedia